Signs
I’ve completed the first pass at signs. They’re shown when you clear
blocks and score combos / chains. Animation was a bitch to figure out,
and I feel like I had to settle for a really tacky approach where signs
sit on an Idle state and are sampled 30 times per second for a trigger
to rise. Ideally, there should be a way to just trigger the rise
without having to poll.